AutoTrackingRewrite: Rewrite __main__ completely!
This is a full rewrite with the goal of automatically tracking every
module within the target module. It does this without the need for the
library in question to use import_tracker at all. It still uses a lot of
the clever bits around spaning subprocesses and using lazy module
initialization to avoid attributing downstream imports to "sibling" modules
but all of that logic is local to __main__.