AutoTrackingRewrite: Eeeeeeven more bug fixes!
Determining when to finalize a deferred attribute is really hard because
while the entries in sys.modules can be deleted to force re-imports, the
attributes on non-lazy modules cannot be deleted without breaking the
from X import Y syntax.

To get around this, we need to actually backtrace the references to a given
lazy module and see if we find the target module upstream.