StackTrackingDeferredSiblings: Partial WIP towards supporting deferred sibling modules in stack tracing
This gets _really_ gnarly. The problem is that when a sibling module to the
module being tracked is imported by the common parent module, it "succeeds"
in importing as a _DeferredModule. If the module in question then attempts
to import the sibling using `from .. import sibling`, that import is
is actually implemented as a `getattr` rather than a full import, so the
stack trace for that import line is not captured.

This branch attempts to solve that by capturing the place where those
sibling imports happen (the place where these _DeferredModules are
triggerd to perform their import) and faking the stack trace entry. There
are several shortcomings with this current approach:

1. It's not recursive, so if a third sibling (submod3) depends on the
   second sibling (submod2) which depends on the first sibling (submod1),
   the stack will not correctly trace submod3 -> submod2 -> submod1

2. The current approach is based on parsing the source code of the file
   where submod2 performs the `from .. import submod1`. This is very error-
   prone since there are numerous ways that this import can happen such as
   multi-line, `import *`, etc... that would not be captured by the current
   parsing logic.