SingleExtras: Handle the single-extras case in parse_requirements
This is not a _full_ solution as it does not cover the case where there's a
single extra and no other non-extras. That case is much harder since the
extra is accounted as a direct child of the parent and thus the inner
tracking will always allocate the extra's requirements to the parent (and
vice versa).

https://github.com/IBM/import-tracker/issues/38