DirectDepsInParseRequirements: Update the import order in the direct dep ambiguous test repo
This is unfortunate, but alog was being attributed to the bar submodule
because it came before bar in the import order. This should also probably
be fixed so that this kind of stuff is order independent!

NOTE: We need to ignore the modules __init__ for iconfig due to import order

https://github.com/IBM/import-tracker/issues/48