LogicOverhaul: First pass at a complete overhaul!
This is a bit of a monolithic commit, but there's no real way to decompose
it. The main change here is that the tracking is now being done post-import
by using `dis` to parse the bytecode of all modules and detect all import
statements. This means that the logic can all live in-process rather than
needing to spawn subprocesses in order to isolate the imports. It also
changes the recursion semantics and removes a bunch of args related to this
from the cli and main function signature.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>