LogicOverhaul: Move more tests over to test_import_tracker
There are some _minor_ output deltas in the direct_dep_ambiguous where
direct_dep_ambigous.bar now depends on alog because it is a direct dep of
the parent module direct_dep_ambiguous. This is more correct and
consistent and avoids one of the places where import order used to
influence the result.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>