LogicOverhaul: Move the test for imports behind LazyModule
BREAKING CHANGE: With the overhaul to use bytecode, lazy deps are fully
invisible since there's no way to detect when a LazyModule triggers its
import without inspecting every __getattr__ invocation in the bytecode!
This is a departure from the results when instrumenting the import
framework where these changes can be detected in sys.modules.
Ultimately, this functionality with LazyModule should probably go away
fully since it's not a particularly userful tool anymore.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>