LogicOverhaul: Remove LazyModule and everything associated with it
This tool was only put into the public API to support a mechanism for
hiding imports from tracking in certain parts of a library. Ultimately,
that paradigm has proven to be a bad one since those hidden imports can
result in fully invisible dependencies if the primary usage modules only
bring in those deps via the LazyModules. The better solution is just to
carefully manage dependencies to only bring in what's needed.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>