SkipOptionalImports: Implement show_optional arg to track_module
This will detect when an import statement is guarded by a try/except and
decorate the output of track_module with "optional": bool to indicate this.
The indication is an aggregate, so if there are three sources for a given
dependency, the "optional" field is only set to True if all three are
optional.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>