IntermediateExtras: Stop iterating the non_extra_union determination if a terminal parent is hit
If a parent of the module in question is a leaf in the extras_modules_tree,
then all children should be considered part of it.

NOTE: This does NOT fix the case where an overlapping parent and child are
both held as extras. This doesn't seem like a particularly legitimate case,
but it could still cause issues if the extras are intentionally nested.

https://github.com/IBM/import-tracker/issues/71

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>