FixIsortVersionDeps: Remove PPA build matrix
These builds were added in response to
https://github.com/IBM/import-tracker/issues/40 which was filed against
a version of the project that inspected the `cmath` package to attempt
to deduce the dirname for dynamically linked standard libs. Since 3.x,
this is not done anymore, so these tests are not needed. Furthermore,
the builds have become out-of-date and would be costly to maintain.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>